fixes bug 1171160 - Splitting up classes in main/tests/test_views.py into different files
